Pass result of "clobber path" prompt to caller, stop task
The end user would be prompted if a file is going to be overwritten,
but homemaker would still attempt to do so no matter the answer.

Move the retry-functionality of cleanPath inside this function so that
the function itself can return a boolean to indicate if the calling task
is free to continue. If the path is not clean then homemaker can just
move on to the next file or next task.